# Small filter efficiency improvements