@KarateBrot 2021-05-17
@ctzsnooze

> I've not really understood why during each Apply we do all these divisions etc...
1 / ( 2 * 1.961459177f * M_PIf * f_cut)
when we could have a constant and one division, ie:
0.08114109381329f / f_cut

Good argument, didn't bother looking at compressing it further. It's possible that the compiler already does this kind of stuff in the background when optimizing for speed but I don't know that for sure.