@McGiverGim 2021-05-17
Yes, the compiler will transform all the constants operations into one, so you DON'T need to compress it further, and the code is clearer.