@KarateBrot 2021-05-19
- Okay, I reverted the stuff that was not dependent on runtime.
- I took the opportunity to clean up some whitespaces and braces.
- This PR will be the base for a further improvement of biquads (notches) in general and the RPM filter

**This PR is flight tested. The biquad still works as expected.**