@TehllamaFPV 2021-05-20
Additional flight testing - from 2.5" to 7" with a dozen race quads in between (RPM notch Q's from 333-1000), still works fine, no practical change to performance of the RPM or dynamic notches in gyro_scaled to gyro performance.